adding range adjustment

+++ include/TBmonit.h
@@ -39,8 +39,8 @@ private:
   std::vector<std::string> fastList_;
   std::vector<std::string> waveList_;
 
-  double DistMaxADC_;
-  double DistBinADC_;
+  float DistMaxADC_;
+  int DistBinADC_;
 
   std::string outputName_;
   std::string mappingpath_;
